tests, flags: Extract flags handling to flags.go
