+++ tests/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "config.go",
+        "flags.go",
         "test.go",
         "utils.go",
     ],
